add mongo db by Moses

+++ src/main/scala/de/htwg/se/uno/aview/Tui.scala
@@ -57,13 +57,13 @@ class Tui(controller: ControllerInterface) extends Reactor {
         Success("Valid Command: " + input)
       case "ld" =>
         if wf.length > 1 then
-          controller.load(wf(2).toInt)
+          controller.load(wf(2))
         else
           controller.load()
         Success("Valid Command: " + input)
       case "dl" =>
         if wf.length > 1 then
-          controller.delete(wf(2).toInt)
+          controller.delete(wf(2))
         else
           controller.delete()
         Success("Valid Command: " + input)
